Add ABTest component
This adds an ABTest package to be used for WCPay experiments.

The package adds a component that accepts an experment name, control,
experiment, and loading components and selects between the two groups
randomly. The package also takes care of persisting the choice in both
localhost and the backend so future visits see consistent results.